privde graph with vertex data
